Split off Agent-related classes and Concept
